Refactor manifest code
